fixed orderbook, spread, floating issue. Limit amount to only integer not float. Limit price to only 2 floating point

+++ trading-engine/engine/order.go
@@ -9,15 +9,15 @@ import (
 var orderType = [3]string{"limit", "market", "stop"}
 
 type Order struct {
-	Amount uint64 `json:"amount"`
-	Price  uint64 `json:"price"`
-	ID     uint64 `json:"id"`
-	Side   int8   `json:"side"`
-	Type   int8   `json:"type"`
+	Amount uint64  `json:"amount"`
+	Price  float64 `json:"price"`
+	ID     uint64  `json:"id"`
+	Side   int8    `json:"side"`
+	Type   int8    `json:"type"`
 }
 
 func (order Order) String() string {
-	return fmt.Sprintf("{ID:%d, Amount:%d, Price:%d, Type:%s} ", order.ID, order.Amount, order.Price, orderType[order.Type])
+	return fmt.Sprintf("{ID:%d, Amount:%d, Price:%f, Type:%s} ", order.ID, order.Amount, order.Price, orderType[order.Type])
 }
 
 func (order *Order) FromJSON(msg []byte) error {
